Moved IVector into main Processor namespace.
