+++ src/bootstrap.php
@@ -46,6 +46,7 @@ const CLASS_MAP = [
   'ABadCafe\\G8PHPhousand\\TestHarness\\Assembler\\Vasmm68k' => '/TestHarness/Assembler/Vasmm68k.php',
   'ABadCafe\\G8PHPhousand\\Processor\\IOpcode' => '/Processor/IOpcode.php',
   'ABadCafe\\G8PHPhousand\\Processor\\TCache' => '/Processor/TCache.php',
+  'ABadCafe\\G8PHPhousand\\Processor\\IVector' => '/Processor/IVector.php',
   'ABadCafe\\G8PHPhousand\\Processor\\IOpcodeMSB' => '/Processor/IOpcodeMSB.php',
   'ABadCafe\\G8PHPhousand\\Processor\\IRegister' => '/Processor/IRegister.php',
   'ABadCafe\\G8PHPhousand\\Processor\\TAddressUnit' => '/Processor/TAddressUnit.php',
@@ -104,7 +105,6 @@ const CLASS_MAP = [
   'ABadCafe\\G8PHPhousand\\Processor\\Opcode\\Template\\Params' => '/Processor/Opcode/Template/Params.php',
   'ABadCafe\\G8PHPhousand\\Processor\\Opcode\\Template\\TGenerator' => '/Processor/Opcode/Template/TGenerator.php',
   'ABadCafe\\G8PHPhousand\\Processor\\Fault\\TProcess' => '/Processor/Fault/TProcess.php',
-  'ABadCafe\\G8PHPhousand\\Processor\\Fault\\IVector' => '/Processor/Fault/IVector.php',
   'ABadCafe\\G8PHPhousand\\Processor\\Fault\\IFunctionCode' => '/Processor/Fault/IFunctionCode.php',
   'ABadCafe\\G8PHPhousand\\Processor\\Fault\\ISpecialStatusWord' => '/Processor/Fault/ISpecialStatusWord.php',
   'ABadCafe\\G8PHPhousand\\Processor\\Fault\\Address' => '/Processor/Fault/Address.php',
